pyproject.toml
[project]
name = "pydantic-ai-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asyncio>=3.4",
    "pydantic-ai[logfire]>=0.0.11",
    "python-dotenv>=1.0",
    "ruff>=0.8",
]
